+++ lib/src/types/multiple_options_details.dart
@@ -1,0 +1,6 @@
+/// A record containing the [value], the [title] and an optional [subtitle] of an option of a [SettingMultipleOptionsTile].
+typedef MultipleOptionsDetails = ({
+  Object value,
+  String title,
+  String? subtitle,
+});