+++ CHANGELOG.md
@@ -5,9 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 3.0.0
+
+### BREAKING CHANGES
+
+- Update `SettingMultipleOptionsTile.detailed` so that the `options` parameter is a list of `MultipleOptionsDetails` record
+
+### Added
+
+- Minimum number of selected options for `SettingMultipleOptionsTile`
+
 ## 2.0.4
 
-Update example
+- Update example
 
 ## 2.0.3